oops

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Pixel.java
@@ -24,7 +24,7 @@ public class Pixel {
     }
 
 
-    public int[] getHSV(){
+    public double[] getHSV(){
         return HSV;
     }
 }
